requirements/dev.txt
alembic==0.8.6
beautifulsoup4==4.4.1
blinker==1.4
cssmin==0.2.0
factory-boy==2.7.0
fake-factory==0.5.7
Flask==0.10.1
Flask-Assets==0.11
Flask-DebugToolbar==0.10.0
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
psycopg2==2.6.1
py==1.4.31
pytest==2.9.1
python-dateutil==2.5.3
python-editor==1.0
six==1.10.0
SQLAlchemy==1.0.12
waitress==0.9.0
webassets==0.11.1
WebOb==1.6.0
WebTest==2.0.21
Werkzeug==0.11.8
WTForms==2.1
